windowmanager: fix crash during workspace switching
when a window is hidden e.g. on focus out it will still appear in the
compositor's window actors list, however its monitor field is already
unset. This will lead to a crash if an unmanaged window is in the window
actors list and we call meta_window_get_monitor() on its metawindow.