Changed type of workspace_view
I assume it is fine to change the type from ActivatableComponent to
DeepinMultitaskingView because from a semantic point of view
workspace_view is always going to be some sort of "multitasking view".

If abstraction is desired, I propose to create an interface for
different types of "multitasking view" plugins, which contains
is_toggling(). Alternatively it could be added to ActivatableComponent
in terms of an abstract function. However, this would require
WindowOverview to be updated as well which would currently be of no use.