Only perform workspace switch when not toggling
Workspace switching is inactive as long as workspace_view is toggling.
This makes sure that the animation does not look broken.

Alternatively we could simply delay the call to switch_to_next_workspace
instead of completely blocking it (using either Timeout or by binding
a callback function to a "toggle finished" signal).